refactor: use mono repo structure
