--- .stylelintrc
@@ -1,13 +1,0 @@
-{
-  "processors": [
-    "stylelint-processor-styled-components"
-  ],
-  "extends": [
-    "stylelint-config-recommended",
-    "stylelint-config-styled-components"
-  ],
-  "rules": {
-    "font-family-no-missing-generic-family-keyword": null,
-    "no-descending-specificity": null
-  }
-}
