+++ packages/backend/favicon.ico
