+++ packages/bot/.env.example
