+++ packages/bot/README.md
