+++ packages/bot/package.json
