+++ packages/bot/src/commands/General/ping.js
