+++ packages/bot/src/config.js
