+++ packages/bot/src/index.js
