+++ packages/bot/src/lib/constants.js
