+++ packages/bot/src/lib/setup.js
