+++ packages/bot/src/listeners/commands/commandDenied.js
