+++ packages/bot/src/listeners/ready.js
