+++ packages/bot/src/preconditions/OwnerOnly.js
